LPS-167061 Extract all submition and validation logic to the new EditObjectField parent component
